feat: update authentication endpoints to use POST method and enhance error handling

+++ spring-security-test/src/main/java/com/pnu/springsecuritytest/entity/type/AuthProviderId.java
@@ -20,5 +20,5 @@ public class AuthProviderId implements Serializable {
     private UUID userId;
 
     @Column(name = "provider_id")
-    private Long providerId;
+    private Integer providerId;
 }
